Refs #1354 Extract cockpit pane app
