+++ src/pollypm/cockpit_apps/__init__.py
@@ -1,0 +1,5 @@
+"""Focused Textual app modules extracted from ``cockpit_ui``."""
+
+from pollypm.cockpit_apps.pane import PollyCockpitPaneApp
+
+__all__ = ["PollyCockpitPaneApp"]
