+++ src/pollypm/cockpit_ui.py
@@ -66,6 +66,7 @@ from pollypm.cockpit_activity import (  # noqa: F401  (re-exported)
     _activity_type_colour,
 )
 from pollypm.cockpit_alerts import _action_view_alerts
+from pollypm.cockpit_apps.pane import PollyCockpitPaneApp  # noqa: F401
 from pollypm.cockpit_inbox import (
     InboxThreadRow,
     build_inbox_thread_rows,
@@ -136,7 +137,6 @@ from pollypm.rejection_feedback import (
 # parameter annotations below (``service: PollyPMService | None``).
 if TYPE_CHECKING:
     from pollypm.service_api import PollyPMService  # noqa: F401
-from pollypm.cockpit import build_cockpit_detail
 from pollypm.cockpit_navigation import (
     InMemoryNavigationStateStore,
     NavigationCommand,
@@ -3709,57 +3709,6 @@ class PollyDashboardApp(App[None]):
             self.config_path,
             client_id="polly-dashboard",
         ).jump_to_inbox()
-
-
-class PollyCockpitPaneApp(App[None]):
-    TITLE = "PollyPM"
-    SUB_TITLE = "Pane"
-    CSS = """
-    Screen {
-        background: #10161b;
-        color: #eef2f4;
-        padding: 1;
-    }
-    #body {
-        border: round #253140;
-        background: #111820;
-        padding: 1 2;
-    }
-    """
-
-    def __init__(self, config_path: Path, kind: str, target: str | None = None) -> None:
-        super().__init__()
-        self.config_path = config_path
-        self.kind = kind
-        self.target = target
-        self.body = Static("", id="body")
-
-    def compose(self) -> ComposeResult:
-        yield self.body
-
-    def on_mount(self) -> None:
-        self._refresh()
-        self.set_interval(5, self._refresh)
-        # #1109 follow-up — TTY-less keystroke bridge.
-        try:
-            from pollypm.cockpit_input_bridge import start_input_bridge
-            bridge_kind = f"pane-{self.kind}"
-            self._input_bridge_handle = start_input_bridge(
-                self, kind=bridge_kind, config_path=self.config_path,
-            )
-        except Exception:  # noqa: BLE001
-            self._input_bridge_handle = None
-
-    def on_unmount(self) -> None:
-        bridge = getattr(self, "_input_bridge_handle", None)
-        if bridge is not None:
-            try:
-                bridge.stop()
-            except Exception:  # noqa: BLE001
-                pass
-
-    def _refresh(self) -> None:
-        self.body.update(build_cockpit_detail(self.config_path, self.kind, self.target))
 
 
 # Re-export the extracted task screen so existing import paths keep working.
